Safari-611.1.10/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-content-before-legend-expected.txt

X legend
Y
legend
X Y

FAIL fieldset content before legend assert_equals: expected 36 but got 56
